Use insecure hackney for testing, refactor to Base
- Paytm’s test server always has an invalid SSL certificate, configure hackney as insecure for it.
- (Production may also have problems, it’s up to the users of this library to configure that)
- Extract the common API functions to a base library
