+++ fixture/vcr_cassettes/api-oauth-send_otp-correct-login-params.json
@@ -4,7 +4,10 @@
       "body": "{\"scope\":\"wallet\",\"responseType\":\"token\",\"phone\":\"7777777777\",\"email\":\"foo@bar.com\",\"clientId\":\"CLIENT_ID\"}",
       "headers": [],
       "method": "post",
-      "options": [],
+      "options": {
+        "recv_timeout": 30000,
+        "insecure": true
+      },
       "request_body": "",
       "url": "https://accounts-uat.paytm.com/signin/otp"
     },
