+++ fixture/vcr_cassettes/api-oauth-validate_token-invalid-token.json
@@ -6,7 +6,10 @@
         "session_token": "foobar"
       },
       "method": "get",
-      "options": [],
+      "options": {
+        "recv_timeout": 30000,
+        "insecure": true
+      },
       "request_body": "",
       "url": "https://accounts-uat.paytm.com/user/details"
     },
