+++ fixture/vcr_cassettes/api-wallet-charge-success.json
@@ -4,7 +4,10 @@
       "body": "{\"scope\":\"wallet\",\"responseType\":\"token\",\"phone\":\"7777777777\",\"email\":\"bar@baz.com\",\"clientId\":\"CLIENT_ID\"}",
       "headers": [],
       "method": "post",
-      "options": [],
+      "options": {
+        "recv_timeout": 30000,
+        "insecure": true
+      },
       "request_body": "",
       "url": "https://accounts-uat.paytm.com/signin/otp"
     },
@@ -32,7 +35,10 @@
         "Authorization": "***"
       },
       "method": "post",
-      "options": [],
+      "options": {
+        "recv_timeout": 30000,
+        "insecure": true
+      },
       "request_body": "",
       "url": "https://accounts-uat.paytm.com/signin/validate/otp"
     },
@@ -60,7 +66,10 @@
         "session_token": "810d0508-c1d9-4d0e-ada1-e6e009eda9ad"
       },
       "method": "get",
-      "options": [],
+      "options": {
+        "recv_timeout": 30000,
+        "insecure": true
+      },
       "request_body": "",
       "url": "https://accounts-uat.paytm.com/user/details"
     },
@@ -87,6 +96,7 @@
       "headers": [],
       "method": "post",
       "options": {
+        "insecure": true,
         "recv_timeout": 30000
       },
       "request_body": "",
